# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Swagger OpenAPI documentation for Laravel API",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/cart/{product_id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete Product from Cart",
                "description": "Remove a product from the authenticated user's cart. Requires a valid Bearer token.",
                "operationId": "e3251b4d3f39406b34a3fb24d7ae3f6e",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product to remove from cart",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 32
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Removed Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 32
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "discounted_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "quantity": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "discount_percent": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "primary_image": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to Remove Product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not enough stock!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token Missing or Invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}